changes to support JSMD import instead of json state storage
