+++ package.json
@@ -67,6 +67,7 @@
     "jest": "^21.2.1",
     "jest-cli": "^22.1.4",
     "neutrino": "^8.1.0",
+    "raw-loader": "^0.5.1",
     "react-addons-test-utils": "^15.6.1",
     "react-test-renderer": "^15.6.1",
     "redux-mock-store": "^1.5.1",
